--- src/App.js
@@ -1,5 +1,0 @@
-class App {
-  async play() {}
-}
-
-export default App;
