+++ src/racingcar/index.js
@@ -1,0 +1,4 @@
+import App from './App.js';
+
+const app = new App();
+app.play();
