+++ src/racingcar/lib/classes/NumberGenerator.js
